Migrate permutations and cartesianProduct, rename * to cartesianProduct
