+++ Sources/DataStructures/ArrayExtensions.swift
@@ -5,8 +5,6 @@
 //  Created by James Bean on 6/29/18.
 //
 
-#warning("Move to Algorithms module")
-
 extension Array {
 
     /// - Returns: `Element` at index if present. Otherwise `nil`.
@@ -54,6 +52,7 @@ extension Array {
 
 extension Array {
 
+    #warning("Move to Algorithms module")
     #warning("Attempt to generalize to RangeReplaceableCollection")
 
     /// - Returns: A two-tuple containing two arrays of `Elements` split at the given `index`,
@@ -130,26 +129,3 @@ extension Array {
         return copy
     }
 }
-
-extension Array {
-
-    #warning("Attempt to generalize to RangeReplaceableCollection")
-
-    public var permutations: [[Element]] {
-
-        // FIXME: Use `ArraySlice` to avoid conversion to `Array`.
-        func permute(_ values: [Element]) -> [[Element]] {
-            guard let (head, tail) = values.destructured else { return [[]] }
-            return permute(Array(tail)).flatMap { injecting(head, into: $0) }
-        }
-
-        return permute(self)
-    }
-}
-
-// FIXME: Use `ArraySlice` to avoid conversion to `Array`.
-internal func injecting <T> (_ value: T, into values: [T]) -> [[T]] {
-    guard let (head, tail) = values.destructured else { return [[value]] }
-    return  [[value] + values] + injecting(value, into: Array(tail)).map { [head] + $0 }
-}
-
